local: store the free blocks locally as a Lvm.Pv.Allocator.t ( = area list)
There's no need for the local machine to store the free blocks in the device
mapper; all we need is to be able to add and remove blocks efficiently.

The remote will consider the free blocks to have been allocated to an LV,
but we don't need to have exactly the same representation.

Signed-off-by: David Scott <dave.scott@citrix.com>