xenvmd: lazily flush the LVM metadata; synchronise with it
We deliberately pause for up to 5s between flushes to allow metadata updates
to be batched.

Unfortunately since this program has a pair of global references:
- myvg
- metadata
which are out-of-sync, in particular 'metadata' is new while 'myvg' is updated
by the global flush, we need to wait for a flush to happen before we can
use Vg_IO.Volume. If we push the out-of-sync code into mirage-block-volume then
this can all be hidden.

Signed-off-by: David Scott <dave.scott@citrix.com>