Prevent the LVs created during format being immediately overwritten
Currently mirage-block-volume is using Random.get to generate ids, which
we use as primary keys. When the volume was formatted the PRNG was in
an initial state, and when xenvmd was started the PRNG was in the same
state. Therefore the sequence of IDs overlapped and volumes went missing.

Related to [mirage/mirage-block-volume#63]

Signed-off-by: David Scott <dave.scott@citrix.com>