local allocator: allocate incrementally
Previously we would try to allocate everything we needed in one chunk.
If the free space isn't available we would block. However xenvmd does
not know we need more space and won't give us more, so we deadlock.

Instead we allocate as much as we can (up to the amount we really need),
expand the volume, thus draining the free pool, which triggers xenvmd
to refill it.

Fix #69

Signed-off-by: David Scott <dave.scott@citrix.com>