test: Do lvchange --offline before shutting down xenvmd
Shutting down xenvmd causes the metadata to be flushed to the disk which we
cannot guarentee to be the case in the use case for `lvchange --offline`.

N.B. this change (intentionally) results in the test failing because we do not
handle the case where a recently created LV has not yet been persisted to disk.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>