model: Add initial model for allocator
This model finds the double-allocation bug caused when xenvmd replays
journalled allocations to the local allocator after crashing.  In
verification mode, the model gets stuck in the Illegal state.

The model does not yet include the fixes for the bug.  It could also
be improved by annotating acceptable end states and adding a never
claim to check that the block process never enters the Illegal state.

Signed-off-by: Euan Harris <euan.harris@citrix.com>