CA-185666: Handle fragmented expands again
The local allocator needs to stat the device to find out where to put the new
extents. If it has less extents than the extend request then it needs to
re-stat the device after partially fulfilling the request to ensure the second
part is put after the first and not overlapping.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>