Add a test to check for problems with FromLVM
In particular, if an error occurs after the ExtendPool message has been
written to the ring but before it has been removed from the journal, the
same blocks will be sent next time the journal is replayed, leading to
double allocation of the same blocks. This test case detects that
problem.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>